Initial commit of issue 2186 for OSX
